San Francisco Hourly Ambulance Call Predictions
Data science and machine learning to predict ambulance calls by hour in San Francisco. This file contains the data preparation as well as a slew of graphs that show the relationships between variables and hourly call volumes. The final Jupyter Notebook shows the predictions at their current state which are within a maximum average error of 3.53 calls/hour and an R squared of 0.4953, which are significant when considering the highly erratic nature of emergency calls. Suggestions and tips are welcomed as this project remains in progress.